Week-4/requirements.txt
# LangChain & LangGraph
langchain>=0.1.0
langchain-aws>=0.2.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.2.0

# AWS Bedrock (LLM + Embeddings)
boto3>=1.34.0

# Vector Store
faiss-cpu>=1.7.4

# Web Search
ddgs>=7.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Google APIs (for MCP server)
google-api-python-client>=2.100.0
google-auth>=2.0.0

# Utilities
python-dotenv>=1.0.0
